+++ client/src/components/auth/register.js
@@ -4,7 +4,7 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { reduxForm, Field } from 'redux-form';
-import { register } from '../actions';
+import { register } from '../../actions';
 import './auth.css';
 
 class SignUp extends Component {
