+++ client/src/index.js
@@ -5,10 +5,10 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware, compose } from 'redux';
 import ReduxThunk from 'redux-thunk';
-import { BrowserRouter, Route } from 'react-router-dom';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import './index.css';
 import App from './App';
-import { SignIn, SignUp, Cards, Card, CardCreate, Settings, Billing } from './components';
+import { Login, Register, Cards, Card, CardCreate, Settings, Billing } from './components';
 import RequireAuth from './components/HOC/RequireAuth';
 import reducers from './reducers';
 // import registerServiceWorker from './registerServiceWorker';
@@ -21,16 +21,16 @@ const store = createStore(reducers,
 ReactDOM.render(
   <Provider store={store}>
     <BrowserRouter>
-      <div>
+      <Switch>
         <Route exact path="/" component={App} />
-        <Route path="/SignIn" component={SignIn} />
-        <Route path="/SignUp" component={SignUp} />
-        <Route path="/Cards" component={RequireAuth(Cards)} />
-        <Route path="/Cards/:id" component={RequireAuth(Card)} />
-        <Route path="/Card/Create" component={RequireAuth(CardCreate)} />
-        <Route path="/Settings" component={RequireAuth(Settings)} />
-        <Route path="/Billing" component={RequireAuth(Billing)} />
-      </div>
+        <Route path="/login" component={Login} />
+        <Route path="/register" component={Register} />
+        <Route path="/cards" component={RequireAuth(Cards)} />
+        <Route path="/card/create" component={RequireAuth(CardCreate)} />
+        <Route path="/card/:id" component={RequireAuth(Card)} />
+        <Route path="/settings" component={RequireAuth(Settings)} />
+        <Route path="/billing" component={RequireAuth(Billing)} />
+      </Switch>
     </BrowserRouter>
   </Provider>,
   document.getElementById('root'),
